+++ examples/xml_utils.py
@@ -42,9 +42,11 @@ from rpclib.model.primitive import Integer
 from rpclib.model.primitive import Decimal
 from rpclib.model.primitive import DateTime
 from rpclib.model.complex import ComplexModel
+from rpclib.model.complex import XmlAttribute
 
 from rpclib.util.xml import get_schema_documents
 from rpclib.util.xml import get_object_as_xml
+from rpclib.util.xml import get_xml_as_object
 from rpclib.util.xml import get_validation_schema
 
 
@@ -63,6 +65,7 @@ class Foo(ComplexModel):
     b = Integer
     c = Decimal
     d = DateTime
+    e = XmlAttribute(Integer)
 
 
 docs = get_schema_documents([Punk, Foo])
@@ -79,8 +82,10 @@ print("the other namespace %r:" % docs['s0'].attrib['targetNamespace'])
 print(etree.tostring(docs['s0'], pretty_print=True))
 
 
-foo = Foo(a='a', b=1, c=3.4, d=datetime(2011,02,20))
-print(etree.tostring(get_object_as_xml(foo),pretty_print=True))
+foo = Foo(a='a', b=1, c=3.4, d=datetime(2011,02,20),e=5)
+doc = get_object_as_xml(foo)
+print(etree.tostring(doc, pretty_print=True))
+print(get_xml_as_object(Foo, doc))
 
 # See http://lxml.de/validation.html to see what this could be used for.
 print(get_validation_schema([Punk, Foo]))